Introducing ManagedCache
Summary:
Process local caching is problematic in a world with actively tracked dependencies.  The problematic scenario is this:

1. Dependency A requires property P of key K.  The calculation of P(K) requires upstream information U.  The handler for P correctly takes a dependency key, and so marks a dependency between A and U in the calculation of P(K).
1. P(K) is placed in a process local cache
1. Dependency B also requires property P of K.  However, because P(K) is already in the cache, it just uses the cached value without marking a dependency between B and U.
1. U is changed by an incremental update.  On updates, we purge the process local caches.  A is correctly scheduled for re-calculation, but we have no way of knowing that we also need to re-calculate B.

There is no way to know that we should have made a dependency edge between B and U without actively calculating P(K), which defeats the purpose of caching.

Another weakness we'd like to fix is that the existing caches are process-local, which means that if the same query is run on different workers, each one has to re-calculate the same, potentially expensive, query on each worker.  Ideally this cache would be in shared memory, so multiple workers can cooperate.

# Solution
However, now that we're transitioning to a lazy environment model, there actually isn't that big of a difference between a lazily computed environment table and a shared memory cache.  This means that we can strategically set things up in a way that we can re-use our lazy environment infra to make a very effective **managed cache**.  Now, the above scenario goes as following:

1. Dependency A requires property P of key K.  The calculation of P(K) requires upstream information U.  Because P is now it's own table, we mark an edge between P(K) and U.
1. We mark an edge between A and P(K)
1. P(K) is placed in shared memory
1. Dependency B also requires property P of K.  We check the table first, and because it's already present, we just read it from shared memory.
1. We mark an edge between B and P(K)
1. U is changed by an incremental update.  This triggers P(K) to be recomputed, and if it changes as well, we trigger **both** A and B

# Legacy Compatibility
We can make this a no-op in legacy mode by maintaining separate process local caches that we maintain in the old way.  This means that we purge them all on an update, and populate them on gets, just like before.
Note that this is still correct since we aren't relying on dependency edges at all in legacy mode, and instead schedule stuff based on the import graph.  Thus there is no concern about having missing dependency edges, since there are no dependency edges.

Note that things are set up in such a way that full checks use the legacy mode, so won't currently benefit from the worker cooperation aspect.  This is because we set it in legacy mode in full checks to avoid wasting time setting dependency edges.  In a follow up, I can add a separate flag to the config to see whether we're in full check or server mode, and run in a middle ground where we use cooperation, but don't worry about setting dependency edges.

# Caveat

The one big difference between caches and lazy environments, namely that we don't know the set of all keys ahead of time.  With our current environments, we always know the full set of keys, since they usually just correspond to class names or global names, of which we maintain a list.  This is not a problem in prod, but presents issues in debugging.  We maintain an infra for dumping shared memory, which depends on having a complete list of possible keys.  In theory we could maybe maintain a manually created list using the same dependency table/encode/decode table approach we do in dependency tracking to do a parallelized append only data structure.  We can return to this later if this presents a problem.

# Rejected Alternatives

In order to accomplish this goal without storing it in shared memory we would somehow need to know the set of dependencies triggered by the calculation of P(K).  If we used the existing dependency tracking infrastructure to do this, we would be adding edges of P(K) depending on U.  We would then like to be able to find U from P(K) when pulling it from the cache for B, but unfortunately the dependency table infra is only designed to be queried in one direction.  This means that we can get P(K) from U, but not the other way around.  This means that in order to do this, we would need to introduce a whole separate system for these reverse edges.  This is ultimately not much easier than building this stuff out, and doesn't get us the cooperation system.

Reviewed By: dkgi

Differential Revision: D18498884

fbshipit-source-id: 91a5f00a4848b2d8607324f1656e55f2710a438a