EmptyStubEnvironment (Introducing the first managed cache)
Summary:
This diff introduces the first managed cache, for the function AstEnvironment.from_empty_stub.  This function is very expensive and thus has had to be cached.  This was actually the first instance where I realized the problem posed by caching, and thus you can see I manually made dependency tracked reads skip the cache.

With the power of managed caches we can now correctly cache this even in server mode.  This should significantly speed up cold starting a server.

Also note that because of all of the work I did from D18521048 to D18536502, inserting this new "environment" into the sequence of environments had almost no consequences for downstream clients: the only changes needed were to the immediately following environment (AliasEnvironment), and anything that was directly using AstEnvironment.from_empty_stub.  This is because now we basically have a singly-backwardly-linked-list of environments, so insertion just touches the following node, while before some accesses were array-like, so required annoying changes.

Reviewed By: dkgi

Differential Revision: D18543243

fbshipit-source-id: aa4580f22c0ae87311be075aba4c3f94b98017e3