Removing the need for update chaining with update_this_and_all_preceding_environments
Summary:
The last violation of my intended state of only needing to update the following environment when inserting a new environment was update chaining.  Now that we have regularized the update interface, we can have the functor call the previous environments update function, and so on and so on until the "root" environment, which is UnannotatedGlobalEnvironment.

Now we can add as many managed cache environments as we want without having to invade the whole universe each time.

Note: we could in theory expose update_only_this_environment as its typing still enforces all the properties we want, but YAGNI as of now, and we can expose it later if we really want to.

Reviewed By: dkgi

Differential Revision: D18536502

fbshipit-source-id: 5299ff13afd6c861fc965b85a3f2fbad24ab9745