Move configuration generation inside command
Summary:
Moving configuration generation to happen inside the command constructor, as we move towards this constructor being the only place that should be assuming the existence of certain argument fields.

The configuration field must be optional because the Initialize command cannot have a configuration; before we relied on lots of if checks in the preprocessing of arguments in pyre.py to set up whether a configuration existed and whether an analysis directory existed. We can implement the same behavior with inheritance now.

Reviewed By: dkgi

Differential Revision: D18453714

fbshipit-source-id: d37215cc48d4ec5142a28ea903b24d42187e3f30