Move argument-modifying preprocessing into commands
Summary:
Move argument preprocessing into the command constructor.

Most of the test changes are caused by capable terminal defaulting to false, rather than true (since all attributes of a mock pass the `if attribute` check). Since this is now an attribute of the command class rather than something that can be easily overridden by a field on the arguments mock, figured it's easier to just make the tests default to the falsey capable terminal behavior rather than patch this function in every unrelated test.

Reviewed By: dkgi

Differential Revision: D18585513

fbshipit-source-id: 49354e61ba8c33c41b00ee90d5b3ad3208dce3f8