Remove log initialize dependence on argument preprocessing
Summary: Avoid reading and modifying arguments when initializing logging. We also shouldn't put the handler that needs cleanup into the arguments namespace object; let's return it instead. When we move this, switch root, and other preprocessing functions into the command constructor, we can keep the handler that needs cleanup as an attribute of the command class.

Reviewed By: sinancepel

Differential Revision: D18587740

fbshipit-source-id: fad99d68eacb9a7e40ec7f0287dbee7da0d9b086