Keep track of overloads in UnannotatedGlobalEnvironment
Summary:
It turns out that the type checker would report overload-related issues (e.g. missing overload implementation, unmatched overloads, etc.) only after seeing the actual body of the overloads. Therefore, we can't simply drop the overloads on the floor when collecting nested function definitions in `UnannotatedGlobalEnvironment` and then rely on the collected definitions in the type checking phase, which is a must-do for function-level type checking.

This diff changes `UnannotatedGlobalEnvironment.collect_defines` such that overloads are picked up along with the function body. The semantics of existing APIs is left unchanged -- `get_define_body` still have the same behavior when there are multiple implementations for the same function name. But for overloads, instead of simple ignoring them we collect them as well and put them next to the body of the function in the shared memory -- The new data structure that contains (optionally) a define body and the associated overload list is named `FunctionDefinition.t`.

Reviewed By: sinancepel

Differential Revision: D18574249

fbshipit-source-id: 6dd80369053b205208d1ee05062e6383bc98243b