Capture `*args` and `**kwargs` properly
Summary:
For `*args` and `**kwargs`, we were blindly putting the starred-names into the scope stack, which is clearly problematic 😅

This diff corrects that by stripping the stars and also keep track of how many stars it stripped in the scope resolution phase. Then, in the preprocessing phase, we expand the annotations to `List[XXX]` or `Dict[str, XXX]` based on what was annotated (`XXX` becomes `typing.Any` if no annotation is provided).

The annotations could also be expanded directly in the scope resolution phase --  I didn't do it just in case that the scope stack will be used for other clients that do not want/require the expansion.

Reviewed By: dkgi

Differential Revision: D18584899

fbshipit-source-id: 6ba2299b6992d1b0cb07a2e1f9d946fdbc76bcbd