Consolidating the mutually recursive attribute/type order related code into one place
Summary:
We'd like to be able to "manage" the AttributeCache and AnnotationCache.  However, in order to guarantee we wont introduce recursive edges, we need to ensure that the caching is only done at the boundary of the rats nest of mutual recursion, and not within it.  The best way to do this is to consolidate all of these mutually recursive functions into one module and then put the caching at the boundary of that.

This diff just does the consolidation, and does not change the caching behavior.

Note that I also pushed several things that were depended on by the bundle of sadness into the respective environments that the can belong at (some of which can go all the way back to AstEnvironment, who knew).

Reviewed By: dkgi

Differential Revision: D18603155

fbshipit-source-id: dea93eaa2eba77d98dc21be9412e723056caca2a