Capture `self` and `cls` properly
Summary:
Consider examples like this:

```
class Foo:
  x: Optional[int]
  def foo(self) -> None:
    def bar() -> int:
      return self.x  # self is captured here
```

The current implementation will treat `self` as not annotated when it is in the capture list. Since my ultimate goal is to treat unannotated captures in the same way as unannotated parameters (i.e. warn and view as `Any`), we are going to get into troubles with the `self` parameter:
- Warning on unannotated `self` parameter lead to poor user experience.
- Treating `self` as `Any` could hide genuine type checking signals. For example, in the code snippet above, we'll lose a valid unmatched return error if `self` becomes `Any`.

Situation would be the same for `classmethod`, where the first parameter is `cls`.

The solution proposed here is to populate the type of `self`/`cls` during the capture population phase. Specifically, if we (syntactically) detect that a parameter is unannotated `self`/`cls`, we fabricate an annotation (from the name of the parent) for it -- similar to how [this is done](https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/fbcode/tools/pyre/analysis/typeCheck.ml?commit=562c18ded4d9fc5721fbab45cfe942ebdbc9a0ff&lines=812-844) in the type checking phase. To detect `self`/`cls` parameters, we'll need the parameter index from the `Scope.Binding.Kind` interface (D18605340) and the signature of the function where the binding occurs from the `Scope.Scope.Kind` interface (this diff).

Reviewed By: dkgi

Differential Revision: D18605602

fbshipit-source-id: 15b524f790ff975ef198ecffbc192895b55b8ffd