Collect type variables from the parent of nesting functions as well
Summary:
This is a small bug fix for D18400709. In that diff, we try to collect type variables from nesting functions in `State.init`, so examples like this would work:

```
T = TypeVar('T')
def foo(x: T) -> Callable[[T], int]:
    def bar(x: T) -> int:
        return 42
    return bar
```

In D18364187, we also try to collect type variables from nesting class, so examples like this would work:

```
T = TypeVar('T')
class A(Generic[T]):
  def foo(self, x: T) -> T: ...
```

However, we failed to consider that nesting functions can themselves be nested inside a class -- which means that previously examples like this wouldn't work:

```
T = TypeVar('T')
class A(Generic[T]):
  def foo(self, x: T) -> Callable[[T], int]:
    def bar(x: T) -> int:
        return 42
    return bar
```

Fortunately, there's a simple fix for it -- we just need to restructure the code in a way such that nesting class is checked for every level of nesting functions. Note that strictly speaking, since we don't support class definitions that are nested inside function definitions, only the `parent` field of the outermost nesting matters.

Reviewed By: sinancepel

Differential Revision: D18610048

fbshipit-source-id: 320b8fe5b2456965bcb64aaf2a48b0bb4d1da479