Capture define signatures
Summary:
Consider examples like this:

```
def foo() -> None:
  def bar(x: int) -> int: ...
  def baz() -> str:
    reveal_type(bar)  # We expect this to be `typing.Callable[[Named(x, int)], int]`
    return bar(42)
```

In function-level type checking, we would expect that the revealed type preserves everything we know about `bar()`. Unfortunately, this is not the case today since we did not keep any annotation for `bar()` when computing the capture set of `baz()`. Additionally, not having an annotation for captured functions is very undesirable for the same reasons that are stated in D18605602:
- Unannotated captures are supposed to be warned and treated as `Any`
- Warning on captured defines, even if the parameters/returns are annotated, would lead to very confusing/frustrating user experience.
- Treating captured defines as having `Any` type may hide genuine type checking signals. For example, in the code snippet above, we'll lose a valid unmatched return error if `bar` becomes Any.

Now, although the problem itself appears to be similar, the solution here has to be a bit different from what I had there in D18605602. This is because fabricating a callable type using standard `typing.Callable` syntax is far more challenging than fabricating an annotation for `self`/`cls`:
- `typing.Callable` does not capture all info on a given define signature -- for example, it won't keep track of things like  parameter names, keyword-only status, variadic usage, etc.
- Our `Type.Callable` does keep track of all info, but creating it heavily relies on the presence of an annotation parser, which is not available in the preprocessing phase.
- We need to duplicate the logic of `AnnotatedCallable.create` at syntactical level, which is not only verbose but also tricky. For example, we perform subtle return annotation transformation based on the `async` flag, `generator` flag, and the presence of certain decorators on the signature -- do we really want to implement the same logic again?

Those concerns lead to the decision presented in this diff -- instead of fabricating annotations in the preprocessing phase, we just grab the entire signature and put it into the capture list, and postpone the translation from define signature to callable type to the type checking phase. This requires us to change the type of `Define.Capture` -- it now has to distinguish between captured variables, which has an annotation attached, and the captured functions, which has a signature attached.

Reviewed By: dkgi

Differential Revision: D18621304

fbshipit-source-id: 057190c4167d80d34155502763445f1178d34517