Inline `Binding.annotation` into `Binding.Kind`
Summary:
The truth is, not all kinds of bindings can be annotated. In fact, only assignment target and exception handler target can syntactically get an annotation.

Therefore, it doesn't make a lot of sense to keep an `annotation` field for all bindings. Let's inline the `annotation` field to `Binding.Kind` so that only bindings that can be annotated contains an annotation field. Those which cannot don't get it.

We perform the refactoring here to allow heterogenous values to be obtained from `Binding.Kind`, which prepares for D18621304.

Reviewed By: dkgi

Differential Revision: D18621297

fbshipit-source-id: eab7861d5b33e34e14f706e01fa0eec1aa6437e7