Count both property getter and setter in UnannoatedGlobalEnvironment.collect_defines
Summary:
Property getter function and setter function share the same name -- who would have thought about that :(

Functions with duplicated names are problematic to us, because we only keep the body of one of them for the type checking phase. Fortunately, we had this exception in D18574249 where if the duplicate name comes from overloads the bodies will be preserved. Now let's extend that exception to properties.

Reviewed By: mrkmndz

Differential Revision: D18641781

fbshipit-source-id: e5d5d9f6ead572a578f82dbfee31f8db85866e3f