Clarify naming of switch_root
Summary:
Know this has been a source of confusion in the past... this function really just looks upwards for the project configuration. Renaming to make it more parallel with the local configuration finder function, which does the same thing except doesn't change the cwd.

Could potentially take the directory switch out of this function entirely, but for now tried to make it more self documenting by giving it the option to change the directory or not with a default param. Not sure if this is best design.

Reviewed By: dkgi

Differential Revision: D18610731

fbshipit-source-id: fcb8ee7d500ca106eea6d858426680b86bb8b50e