infer models for dataclass __init__
Summary:
Context: Pyre used to generate actual __init__ functions for dataclasses in the preprocessing phase that we could use to infer the taint.
However, this changed in the last month, and we were treating dataclass constructors as obscure and losing the precise taint. (D17064213 killed plugins).

Work around this by special casing the taint of dataclass constructors in the init.

It seemed reasonable to special case the taint of dataclasses here, as their attributes are special cased in the type checker as well.

Reviewed By: dkgi

Differential Revision: D18634006

fbshipit-source-id: 383b96299353bc555de519b20c532d6429941e32