Capture `self` and `cls` properly, attempt II
Summary:
My previous attempt at this (D18605602) reached a flawed solution. This diff tries to fix that.

To recap, what we want to do here is to correctly annotate the captured `self` parameter in an inner function `Foo.foo.bar()`:
```
class Foo:
  x: Optional[int]
  def foo(self) -> None:
    def bar() -> int:
      return self.x  # self is captured here
```

My old solution in D18605602 is to fabricate a syntactical annotation for `self` in `bar()`. It worked great for simple classes, but not so for generic classes:
```
T = TypeVar("T")
class Foo(Generic[T]):
  def foo(self) -> None:
    def bar() -> None:
      reveal_type(self)  # Should be `Foo[T]`, not just `Foo`
```

Missing the parametric annotation there is problematic because in strict mode, we are going to emit a spurious `InvalidTypeParameter` error there :(

As it is tricky to grab the generic parameters of the parent class at preprocessing stage, let's stop fabricating annotations altogether and mimic what was done in D18621304 instead -- postponing the annotation of `self`/`cls` parameter to the type checking phase. This requires us to extend `Define.Capture.Kind` again, keeping track of the name of the parent class and whether it's `self` or `cls`.

Reviewed By: dkgi

Differential Revision: D18652329

fbshipit-source-id: ed0b9f869844d2f13cf990fda1514cebec88cddd