Fix expression show for stubs
Summary:
Looking for some feedback on this approach.

I wanted to avoid adding a whole new `pp` method - since most of the code would just be duplicated. This allows us to add special cases to the show method without overriding the built in signatures for the derived show methods.

If anyone else has a better suggestion, happy to implement it.

Reviewed By: dkgi

Differential Revision: D18625178

fbshipit-source-id: fcd22c9ed8f3c10d1820b7b6de13f3c353cb47e5