Send `hg diff` and commit hash on file save.
Summary:
Problem: Currently, when we send an Update message upon file save, we don't include the root directory of the configuration or the hash of the current commit. This makes it harder to construct specifications since the user may have rebased before a file change or may have changed other files.

We now send the `hg diff` output as content, no longer send the individual file path, and send the project directory path.

Reviewed By: grievejia

Differential Revision: D18584179

fbshipit-source-id: 7011cbdbcf75e608b637fa689ffa83b5ccf0bf77