Changing the interface of attribute_table to operate on names instead of classes
Summary:
In the next diff I'm going to "manage" the attribute cache, which means that we're going to put it into a shared memory table, keyed by the arguments of the function that produces it.  This implies that we're going to want to pare down the size of those arguments as much as possible.  As of now, that argument would need to include an entire class summary, which while not as big as entire class definition, would be orders of magnitude bigger than any of the other shared memory keys we already have.  Furthermore, with the keying on definition approach, class summary changes would directly trigger type checking, rather than just triggering the attributes table, which will only trigger type checking if the attributes actually change.

To avoid this, this diff takes advantage of the fact that most calls to attributes are immediately preceded by direct calls to UnannotatedGlobalEnvironment.get_class_summary or indirect calls to it through UnannotatedGlobalEnvironment.resolve_class, which calls get_class_summary.  This means that we can instead put the fetch inside of the attributes_table, and get much the same effect.

Reviewed By: grievejia

Differential Revision: D18647717

fbshipit-source-id: 2c8abc9cccbbd59609a866af8bfd7f8b55b4f7ba