Use profile log as the source of truth for full check time
Summary:
The differences in how we get the metrics between full and incremental check really gets me -- we grab the full check time by measuring on the Python side but relies on the OCaml-side logging for the incremental check time.

Let's make it consistent by having both of them based on the OCaml-side logging.

Reviewed By: sinancepel

Differential Revision: D18694627

fbshipit-source-id: c04d111d6751f7248e4a37544fc6d7eeb5df21b1