Rewrite `TypeCoverageRequest` handling to use `LookupCache` instead of coverage shared memory
Summary: Computing coverage metrics is rather cheap -- storing coverage numbers in shared memory just for `TypeCoverageRequest` seems like a wrong tradeoff to me. It makes much more sense to recompute coverage numbers on-the-fly when it gets queried -- we've already done this for the `CoverageInFile` query. Tthis way we can avoid all the sharemem dependency tracking headache and easily gets the guarantee that coverage numbers are always up-to-date.

Reviewed By: sinancepel

Differential Revision: D18702261

fbshipit-source-id: 6f1b263981b24000b15b5c8b23daf396262e38d5