only schedule server integration test with the latest client
Summary: We were scheduling the server integration test with old clients, causing us to not test the client that was actually built. Only keep the make server_integration_test call in legocastle_build.sh, as that's the one that overrides the client (and verify that we actually schedule this test).

Reviewed By: grievejia

Differential Revision: D18716210

fbshipit-source-id: 5d0598d5593a99865dc4c0cd127f1ffca4a7c325