expose ReadOnly interfaces for environments in GlobalResolution and use these
Summary: The problem this diff solves is that in the current state of affairs, the consumer has to care about which environment depends on which other one during construction. This is an important detail when building the environment, but for the ReadOnly interface after the environment has been built, it's a nuisance. Use the strategy of exposing the APIs we already have in global_resolution.ml and using them to clean up some client code.

Reviewed By: shannonzhu

Differential Revision: D18726555

fbshipit-source-id: 1f4075860e5e85eea805f24ee578b50e58a1299a