Decouple the API of additional checks from the API of TypeCheck
Summary: We used to have the constraint that `TypeCheck` module interface needs to conform to `Check.Signature`. However, after D18697230 I realized that this doesn't have to be the case any more, as the type check no longer share most of its code path with any other check. The only remaining shared code path between type check and the additional checks is `Check.run_check`, which this diff eliminates.

Reviewed By: sinancepel

Differential Revision: D18722269

fbshipit-source-id: 7eaabe4823aea5a70289468cd007dc68797048f0