support `reveal_type()` in the taint analysis as well
Summary:
The way we implement reveal_type is by type errors. This is important in the type checking
fixpoint, as we don't want to show incomplete types to users. However, the taint analysis doesn't
show the type checking errors, and the type information is most useful when you're going through the taint, so add support the Log.dump type information in the taint analysis as well.

Reviewed By: shannonzhu

Differential Revision: D18734191

fbshipit-source-id: 23affe306bba5bf6d399cb2740b14dc621f5f3eb