Remove `DependentDefine` query
Summary:
Almost forgot that I added this when the fine grained dependency work just got started.

It's been a while since then. I envisioned that we'll need to maintain our own infra for dependency table, but we ended up relying on `hack_parallel`. I also envisioned that we'll have modules as dependency units for a while, but we ended up depending on class names / function names directly.

... which means that the infra has changed so much that it's no longer useful to maintain this query. I can see that making queries about dependency units might be useful in the future, but even if it is, it certainly won't take the same form as `TypeQuery.DependentDefine`. Kill the query for now to reduce maintenance burden.

Reviewed By: dkgi

Differential Revision: D18734511

fbshipit-source-id: 2b64f28448bf7055f46561c68fa268fc71cbe315