Fix flasky upgrade test
Summary:
The problematic test here seems to shell out to `arc lint`, which might take a while (several seconds) to return. This leads to frequent timeout of this test, as shown in the linked task.

From the context it looks like we don't really care about what the linter returns here. So to patch the issue I'll just mock the return value to 0.

Reviewed By: sinancepel

Differential Revision: D18752288

fbshipit-source-id: 803bcc9ad0b6512d55087fd20a6e6119918bcc23