Kill `open_document` flag on `analyze_sources`
Summary:
The reason why `open_documents` needs to be passed to `analyze_sources` in the first place is to mess up with the `store_type_check_resolution` configuration for autocomplete

With D18752779, there's an alternative solution to the local annotation map not available issue. Besides, auto complete is not enabled and not actively maintained -- let's just remove the flag for now.

Reviewed By: dkgi

Differential Revision: D18774607

fbshipit-source-id: c18ec1a717cef87f27692d9d208c65d1528065de