Kill `Analysis.Check.analyze_and_postprocess`
Summary:
Context
======

The end goal I want to achieve here is to migrate from a module-based API to a function-based API for type checking, which is the last blocker for function-level dependency.

I identified three different module-based type checking APIs (i.e. asking for the type errors of a single module or a list of modules) in our codebase:
- `Analysis.Check.analyze_and_postprocess` (5 callsites)
- `Analysis.Check.analyze_sources` (2 callsites)
- `TypeCheck.run` (6 callsites)

If we count how many additional indirect/transitive callers we have, we'll quickly find that migrating everything in one diff is not ideal. Instead, the strategy I'm going to take here is to split the effort across several diffs, gradually chipping away the number of callsites that need to be migrated, until the final migration becomes manageable.

This Diff
======

Let's start with the simple one, `analyze_and_postprocess`.

In test, we mostly use it to obtain postprocessed errors and to initialize server states. As those tests are all migrated to use `ScratchProject`, it is straightforward to lift the said logic into a common function `ScratchProject.build_type_environment_and_postprocess`.

Once all test usages are eliminated, there are only 2 remaining callsites left on prod. We can simply inline the function body into those callsites and get rid of `analyze_and_postprocess` completely.

Reviewed By: sinancepel

Differential Revision: D18736896

fbshipit-source-id: c73c9332507037645c62ffdf90f8c909d4f8873a