"Managing" the AttributeCache
Summary:
This diff introduces the second of our managed caches, this time replacing the AttributeCache.  As described in D18498884, unmanaged caches are intrinsically damaging to our dependency tracking efforts, so these two caches in AttributeResolution are major blockers to our launch of fine-grained dependencies.

To take maximum advantage of the cache, we have be a little tricky and make another "poor man's object" in order to do a kind of OOP open recursion.  This allows our implementation to use an uncached version of itself both directly and indirectly, while allowing the public version of all of the other functions to be defined in terms of the cached version.

You might expect for there to be a perf regression here because we no longer cache the recursive calls to attribute_table from inside of attribute_table.  I suspected initially there would be a problem with having to recalculate the attributes of `object`, for example, on every transitive lookup.  However, we actually already do that, because we key the existing cache on instantiated type.  This means there is no regression from that.  In theory we could maybe get some future wins from selectively ignoring the transitive key?

Reviewed By: grievejia

Differential Revision: D18652652

fbshipit-source-id: d688de360ac7236ee72313e7a9b015f90a129c50