Auto approve pyre bot upgrade diffs
Summary:
Unfortunately it appears as though grabbing the diff # from the output of jf submit is the most direct way of getting this; seems like it's how bot approval is done elsewhere too.

Also, bot approval will fail if this script is run locally. I was weighing whether to make `--bot` a flag, which would allow me to make the script fail or error/warn when bot approval doesn't work. But seemed like maybe this is unnecessary; we can easily tell if bot approval is not working for some reason on PyreBot Jr. diffs without adding more options.

Reviewed By: sinancepel

Differential Revision: D18778337

fbshipit-source-id: 3e7ccd31a44e3d76617e1224c047c12966d31408