Initialize logging once when original command is created
Summary:
Thanks Maggie for catching this issue!

We should grab what we need from the command we created and do the log initialization where the command is created, rather than in the constructor of the base command where this could be duplicated in commands that create other commands in their _run logic.

This still ensures log.initialize doesn't rely on modifying arguments / argument preprocessing isn't needed before initialize, so shouldn't be any drawbacks here.

Reviewed By: sinancepel

Differential Revision: D18802373

fbshipit-source-id: a38c2d9cfcc206f25a5143c7de078d5a6f62b66c