Remove impossible parameters to Initialize constructor
Summary:
By definition, Initialize does not reference a configuration or analysis directory. In the past, we special cased this in pyre.py; in the client command refactor, we were able to differentiate Initialize to avoid special casing by modifying args and also to soundly type the configuration attribute in all other commands as a non-optional.

As a result, the Initialize command does not need to support these parameters in its constructor.

Reviewed By: grievejia

Differential Revision: D18802372

fbshipit-source-id: c8798f043006f27f6d95564e63a0786de24aad98