Move working directory logic out of command initialization
Summary:
We can't call `os.getcwd` inside the command constructor because we wind up changing the working directory after we create the command. This means any commands that create other commands in their _run logic will end up with an inaccurate original directory due to a second `os.getcwd` call.

This is gross and the real solution should be to do another pretty major client refactor that avoids changing directories at all, removing the need for the derived attributes (formerly arguments) current_directory and original_directory. There is a task for this: T57959968 and a TODO in pyre.py.

In the meantime, let's "revert" this piece of logic back to what it used to be before the client refactor, except we'll pass in the result of `os.getcwd` directly into the constructor rather than modifying the parsed arguments and tacking this on as a "fake" argument.

Reviewed By: grievejia

Differential Revision: D18802374

fbshipit-source-id: 518efb673e0a1f409e52d3e48d0f1b4f7e416da1