Managing the AnnotationCache
Summary:
With the benefit of the ManagedCache infrastructure and the open_recurser magic, adding in the final managed cache was incredibly straightforward.

I chose to lay it out AnnotationCache -> AttributeCache rather than the other way around because every attributes call needs to parse annotations, but only annotation parses that have a generic parameter whose upper bound is a protocol should need to call attribute cache.  In theory though, switching them around would be as easy as flip-flopping the `produce_value`s

Reviewed By: grievejia

Differential Revision: D18668245

fbshipit-source-id: dc6b9a4d7357cb279df581a18f1052ee25904dc1