Do not store function bodies in external sources
Summary:
Turns out that one more issue came up when I try to switch from module-level `TypeEnvironment` to function-level `TypeEnvironment`.

A module will *not* be type checked when it is determined to be external (i.e. `source_path.is_external` flag is `true`).  This external module filtering happens in the type checking phase, and it works great now because each module knows whether it is external or not. Problem is, we don't store the same info for function definitions -- meaning that each function does *not* know whether it is external or not, which makes it impossible to correctly do incremental rechecks when the external-ness of a function changes under the hood.

There are multiple ways to approach this problem. The most straightforward way is to add a `is_external` flag to `UnannotatedGlobalEnvironment.FunctionDefinitions`, and make sure the type checker consult it before start checking the function. This is probably also the least disruptive solution I can think of, but it does have its downside: wasted sharedmem space.

The wasted space issue is largely pre-existing as well: if a function is never going to be type checked, why are we storing its body in sharedmem in the first place? Considering that we are already storing the AST of external sources twice in the parser, it seems even more wasteful to store those bodies one more time. This wastefulness motivates my solution here: as the type checker grab function bodies from `UnannotatedEnvironment`, what if we just refuse to store the bodies there to begin with?

Reviewed By: dkgi

Differential Revision: D18793222

fbshipit-source-id: b35619c078c515fde9ffb15f57949945628621a5