Resolve literal vs `Optional` types for List, Dict, and Set.
Summary:
This was a known problem: When the input was a list literal and the expected type was optional, we did not weaken the type when variance was involved.

I'm not handling all the cases because the general solution for weakening literals requires whole-function constraint-solving, which we don't currently do.

Deleted obsolete comment.

Reviewed By: sinancepel

Differential Revision: D18812666

fbshipit-source-id: b8fad06c6776cf70df4889b38e005e95fcbc93d0