Turn the keys of `TypeEnvironment` from module names to function names
Summary:
With all the preparation work checked in, we are now ready to switch to function-level `TypeEnvironment` (i.e. both type errors and local annotation maps are internally stored by `TypeEnvironment` per-function, as opposed to per-module).

The immediate benefit of it is that to get the local annotation map for a given function, we no longer need to grab the list of annotation maps for the entire module and then perform a linear scan -- the scan phase can be skipped and the lookup becomes `O(1)` instead. We also no longer need to supply the module name when making the local annotation map query, which is not super important but nice to have. Finally, this change is a pre-requisite us for fine-grained dependencies -- with module-level `TypeEnvironment`, we have to invalidate the entry for and then recheck the entire module even if only a small subset of functions gets updated in an incremental recheck. In other words, function-level dependencies can only be effectively leveraged when the `TypeEnvironment` tables are also organized by functions.

The actual changes are fairly straightforward: other than tweaking the APIs of `TypeEnvironment` to cope with the new meaning of keys, we simply grep for all uses of `TypeEnvironment.set_XXX` and `TypeEnvironment.get_XXX`, and make sure that the clients' views of what are stored in those tables are adjusted accordingly.

Reviewed By: dkgi

Differential Revision: D18801644

fbshipit-source-id: 006b42bd2c1231d0a8d1c09e602b72b2449549cf