Correct class definitions of TypedDictionary and NonTotalTypedDictionary and use them correctly
Summary: We use the class definitions of TypedDictionary and NonTotalTypedDictionary in order to correctly interface between typed dictionaries and protocols.  Unfortunately those class definitions were a little wrong (we weren't including the "total" special methods in non-total, which are actually common to both) and were applied incorrectly (we used TypedDictionary in both cases, meaning non-totals appeared to not have pop/del/etc).

Reviewed By: dkgi

Differential Revision: D18871556

fbshipit-source-id: 9d434a3732a993ccd39947bd22c18036eaf51d70